requirements.txt
Django==1.11
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
mysqlclient==1.3.12
Pillow==5.1.0
psycopg2==2.6.2
pytz==2018.3
sqlparse==0.2.4
